Only download file if it is not present

+++ vncserver/controller-v3.sh
@@ -114,7 +114,9 @@ if [ -z "${service_vnc_exec}" ] || ! [ -f "${service_vnc_exec}" ]; then
         echo "(date) ERROR: No vncserver or singularity command found"
         exit 1
     fi
-    wget -O ${service_vncserver_sif} https://github.com/parallelworks/interactive_session/raw/vncserver-singularity/downloads/vnc/vncserver.sif
+    if ! [ -f ${service_vncserver_sif} ]; then
+        wget -O ${service_vncserver_sif} https://github.com/parallelworks/interactive_session/raw/vncserver-singularity/downloads/vnc/vncserver.sif
+    fi
     if ! [ -f ${service_vncserver_sif} ]; then
         echo "$(date) ERROR: Failed to download file ${service_vncserver_sif}"
         exit 1
